Update README.md to emphasize async nature of response streaming
In the original example, the response stream was being filled before `myHandler` finished executing. 

Since the non-AWS proxy in `streamifyResponse` awaits on the void promise returns by the provided handler, this means that the lambda would only output values that were fed into the stream imperatively inside the handler function.  Anything that happens asynchronously would not be captured.

The solution to this is to use a promise that only resolves once the potential asynchronous work has been completed.